backend-lambda/requirements.txt
fastapi==0.112.0
mangum==0.17.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
requests==2.32.3

